fixed deleted function

+++ bunny/components/Post.jsx
@@ -10,46 +10,65 @@ import {Link} from 'react-router';
 
 @observer
 export default class Post extends React.Component {
+    /**
+     * Handle the post, tag and edit and delete button
+     */
 
     componentDidMount() {
         this.props.route.posts.pageId = this.props.params.id;
         this.props.route.posts.findPostById();
-        console.log(toJS(this.props.route.posts.post['tag']));
     }
 
-    delete() {
-        axios({
-            method: 'delete',
-            url: '/api/diary/' + this.props.route.posts.pageId + '/',
-            headers: {
-                'Authorization': "JWT " + sessionStorage.getItem('token')
+    deletePost() {  // delete button
+        var _id = this.props.route.posts.pageId;
+
+        $('[data-toggle=confirmation]').confirmation({
+            rootSelector: '[data-toggle=confirmation]',
+            onConfirm: function () {
+                axios({
+                    method: 'delete',
+                    url: '/api/diary/' + _id + '/',
+                    headers: {
+                        'Authorization': "JWT " + sessionStorage.getItem('token')
+                    }
+                }).then(function (data) {
+                    sweetAlert("Saved", "Deleted Successfully", "success");
+                    browserHistory.push('/dashboard/posts/');
+                }).catch(function (error) {
+                    sweetAlert("Warning", "Error in deletion", "error");
+                })
             }
-        })
+        });
     }
 
-    tags() {
+    tags() { // generate tags
         var bunny = [];
-        (this.props.route.posts.post['tag']).map(function (data) {
-            bunny.push(
-                <button className="btn btn-default" role="button" >
-                    #{data}
-                </button>
-            )
-        });
-        return (
-            <div> {bunny} <Link className="btn btn-info"
-                                 to={'/dashboard/posts/' + this.props.route.posts.pageId + '/edit/'}
-                                 role="button">Edit</Link>
-                <button className="btn btn-danger"  role="button" onClick={this.delete()}>Delete</button>
-            </div>
-        );
+        if(this.props.route.posts.post['tag']) {
+            (this.props.route.posts.post['tag']).map(function (data) {
+                bunny.push(
+                    <button className="btn btn-default" role="button" key={Math.random()}>
+                        #{data}
+                    </button>
+                )
+            });
+            return (
+                <div> {bunny} <Link className="btn btn-info"
+                                    to={'/dashboard/posts/' + this.props.route.posts.pageId + '/edit/'}
+                                    role="button">Edit</Link>
+                    <button className="btn btn-danger" data-toggle="confirmation" onClick={this.deletePost()
+                    }>Delete</button>
+                </div>
+            );
+        }
     }
 
     render() {
+        var _title = this.props.route.posts.post['title'];
+
         return (
             <div className="posts">
                 <Helmet
-                    title="Hiren-Diary: Post"
+                    title= {_title}
                     link = {[
                     {"rel": "stylesheet", "type": "text/css", "href": "https://fonts.googleapis.com/css?family=Lora:400,700,400italic,700italic"},
                     {"rel": "stylesheet", "type": "text/css", "href": "https://fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800"},
